integrate assignparser to now use ExpressionUtil functionalities to parse assign statements

+++ Code/source/QueryParser.cpp
@@ -204,7 +204,7 @@ vector<pair<string, string>> QueryParser::splitDeclarations(vector<string> claus
 string QueryParser::validateDeclarations(vector<pair<string, string>> declarations) {
 	// TODO: verify varName with lexical token
 
-	for (int i = 0; i < declarations.size(); i++) {
+	for (size_t i = 0; i < declarations.size(); i++) {
 		if (validVarType.find(declarations[i].first) == validVarType.end()) {
 			return "invalid query";
 		}
